refactor: move enrollment JWT parser out of main.rs
`main.rs` is the service entry point — it shouldn't also be a home for
wire-format-adjacent types. `enrollment.rs` already owns the rest of
the enrollment flow (HTTP call, cert persistence), so `EnrollmentJwt
Claims` and `parse_enrollment_jwt` belong there.

- Move `EnrollmentJwtClaims` struct + `parse_enrollment_jwt` fn from
  `devolutions-agent/src/main.rs` to
  `devolutions-agent/src/enrollment.rs`. Both become `pub` (consumed
  by `main.rs` through the lib→bin boundary).
- Move the two parser-specific unit tests (`parse_enrollment_jwt_
  rejects_malformed`, `parse_enrollment_jwt_requires_gw_url`) into a
  new `#[cfg(test)]` block in `enrollment.rs`, alongside their own
  `make_jwt` helper. The `parse_up_command_args_accepts_enrollment_
  string` test stays in `main.rs` since it's exercising CLI parsing,
  not the JWT decoder itself.
- `main.rs` now does `use devolutions_agent::enrollment::parse_
  enrollment_jwt;` and drops its top-level `base64::Engine` import
  (only the test module still needs it, so the trait use moves into
  the test `mod tests`).
- Drive-by: collapse `anyhow::Result` / `anyhow::bail!` in
  enrollment.rs to the freshly-imported `Result` / `bail` so they are
  consistent across the module.

Tests: 27 lib + 3 bin on devolutions-agent, all green. Zero clippy
warnings on the agent crate. Fmt clean.
